pyproject.toml
[project]
name = "chat-with-your-data-solution-accelerator"
version = "2.0.0"
description = "Chat with your data solution accelerator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-functions>=1.24,<2.0",
    "python-dotenv>=1.2,<2.0",
    "azure-ai-formrecognizer>=3.3,<4.0",
    "azure-storage-blob>=12.28,<13.0",
    "azure-identity>=1.25,<2.0",
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34,<1.0",
    "httpx>=0.28,<1.0",
    "pydantic>=2.11,<3.0",
    "pydantic-settings>=2.7,<3.0",
    "openai>=2.26,<3.0",
    "requests>=2.32,<3.0",
    "tiktoken>=0.12,<1.0",
    "azure-storage-queue>=12.15,<13.0",
    "beautifulsoup4>=4.14,<5.0",
    "fake-useragent>=2.2,<3.0",
    "chardet>=5.2,<6.0",
    "azure-search-documents>=11.6.0b1",
    "azure-ai-contentsafety>=1.0,<2.0",
    "python-docx>=1.2,<2.0",
    "azure-keyvault-secrets>=4.10,<5.0",
    "pandas>=2.3,<3.0",
    "azure-monitor-opentelemetry>=1.6.10,<2.0",
    "opentelemetry-instrumentation-httpx>=0.52b0",
    "pillow>=11.0,<12.0",
    "azure-mgmt-cognitiveservices>=14.1,<15.0",
    "jsonschema>=4.25,<5.0",
    "azure-ai-ml>=1.31,<2.0",
    "rfc3987>=1.3,<2.0",
    # Orchestrators — all co-installed so admin UI can switch at runtime
    "langchain>=0.3,<1.0",
    "langgraph>=0.4,<1.0",
    "langchain-openai>=0.3,<1.0",
    "semantic-kernel>=1.39,<2.0",
    "azure-ai-agents>=1.2.0b3",
    # Database backends — both co-installed so admin UI can switch at runtime
    "azure-cosmos>=4.14,<5.0",
    "asyncpg>=0.31,<1.0",
    "psycopg2-binary>=2.9,<3.0",
    "pgvector>=0.4,<1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0,<10.0",
    "pytest-cov>=7.0,<8.0",
    "flake8>=7.3,<8.0",
    "black>=25.12,<26.0",
    "pre-commit>=4.5,<5.0",
    "pytest-httpserver>=1.1,<2.0",
    "trustme>=1.2,<2.0",
    "jupyter>=1.1,<2.0",
    "pytest-asyncio>=1.2,<2.0",
]

[tool.uv]
# semantic-kernel transitively requires pre-release azure-ai-agents
prerelease = "if-necessary-or-explicit"

[tool.coverage.run]
omit = [
    "**/tests/*",
    "**/dist/*",
    "**/infra/*",
]

[tool.coverage.report]
skip_empty = true
include_namespace_packages = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
